eco32: ld: use generic elf script template
This removes the custom elf32eco32 script template and replace
it with the generic elf template.
In addition to that, 'EMBEDDED=yes' is set to remove headers
from the beginning of the .text session, since we want the
vectors to start at offset 0.
To ensure that the vectors get placed first, a dedicated
.vectors section is created and assigned to
INITIAL_READONLY_SECTIONS.